+++ CHANGELOG.md
@@ -78,8 +78,43 @@
 
 ## Fixed some bugs 🐞
 
+<br>
+<br>
+<br>
+
 # **v1.0.0**
 
 ## ✔️ **New** methods have been added:
 
 - **modulo()** // Get the module of an operation
+
+<br>
+<br>
+<br>
+
+# **v1.1.0**
+
+## ✔️ **New** methods have been added:
+
+- **format()** // Format the currency value with the separators
+  you set in your setup
+
+## ✔️ Now you can get your currency with all the values, only by placing **'full'** in the settings 💯
+
+## ✔️ You can use the **Separators _class_** to configure separators. <br> Now you can import the class and configure them :D
+
+```js
+import Boldor, { Separators } from 'boldor'
+
+const myCashSetup = {
+  currency: 120000.25,
+  precision: 'full',
+  separators: new Separators(['.', ',']), // Or ['.', ',']
+  lang: 'en',
+}
+
+const myCash = new Boldor(myCashSetup)
+
+console.log(myCash.format()) // Output: 120.000,25
+
+```